# backend/docs/swagger.json
{
    "components": {
        "schemas": {
            "APIError": {
                "properties": {
                    "errors": {
                        "description": "detalle de validación, uno por campo",
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "message": {
                        "example": "Recurso no encontrado",
                        "type": "string"
                    },
                    "path": {
                        "example": "/api/v1/meter/getById/M-999",
                        "type": "string"
                    },
                    "status_code": {
                        "example": 404,
                        "type": "integer"
                    },
                    "timestamp": {
                        "example": "2026-09-24T15:00:00Z",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Anomaly": {
                "properties": {
                    "anomaly": {
                        "type": "boolean"
                    },
                    "confidence": {
                        "type": "number"
                    },
                    "detected_at": {
                        "type": "string"
                    },
                    "evidence": {
                        "$ref": "#/components/schemas/Evidence"
                    },
                    "meter_id": {
                        "type": "string"
                    },
                    "narrated_by": {
                        "description": "quién redactó reason y recommended_action: \"engine\" u \"openai:\u003cmodelo\u003e\"",
                        "example": "engine",
                        "type": "string"
                    },
                    "priority": {
                        "description": "1 = revisar primero",
                        "type": "integer"
                    },
                    "reason": {
                        "type": "string"
                    },
                    "recommended_action": {
                        "type": "string"
                    },
                    "severity": {
                        "$ref": "#/components/schemas/Severity"
                    },
                    "type": {
                        "$ref": "#/components/schemas/AnomalyType"
                    }
                },
                "type": "object"
            },
...
            "AnomalyType": {
                "enum": [
                    "REAL_ANOMALY",
                    "EXPLAINABLE_ANOMALY",
                    "FALSE_POSITIVE",
                    "DATA_QUALITY"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "TypeReal",
                    "TypeExplainable",
                    "TypeFalsePositive",
                    "TypeDataQuality"
                ]
            },
            "Check": {
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "strength": {
                        "description": "de 0 a 1: qué tanto se cumple",
                        "type": "number"
                    },
                    "weight": {
                        "description": "cuánto pesa esta señal (los pesos suman 1)",
                        "type": "number"
                    }
                },
                "type": "object"
            },
...
            "ConsumptionChange": {
                "properties": {
                    "actual_kwh": {
                        "description": "lo que consumió de verdad",
                        "type": "number"
                    },
                    "change_pct": {
                        "description": "variación porcentual",
                        "type": "number"
                    },
                    "direction": {
                        "$ref": "#/components/schemas/Direction"
                    },
                    "end": {
                        "type": "string"
                    },
                    "expected_kwh": {
                        "description": "lo que debería haber consumido en ese tramo",
                        "type": "number"
                    },
                    "hours": {
                        "type": "integer"
                    },
                    "max_abs_z": {
                        "description": "la lectura más extrema del tramo",
                        "type": "number"
                    },
                    "ongoing": {
                        "description": "true si sigue activo al final de los datos",
                        "type": "boolean"
                    },
                    "start": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "DailyPoint": {
                "properties": {
                    "baseline_kwh": {
                        "example": 1052.15,
                        "type": "number"
                    },
                    "consumption_kwh": {
                        "example": 1650.2,
                        "type": "number"
                    },
                    "date": {
                        "example": "2026-09-12",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Detail": {
                "properties": {
                    "anomaly": {
                        "$ref": "#/components/schemas/Anomaly",
                        "description": "explicación, acción y evidencia; null si no hay anomalía"
                    },
                    "anomaly_type": {
                        "$ref": "#/components/schemas/AnomalyType"
                    },
                    "baseline_kwh": {
                        "description": "consumo diario normal",
                        "example": 1052.15,
                        "type": "number"
                    },
                    "consumption_kwh": {
                        "description": "consumo de las últimas 24 h",
                        "example": 2207.6,
                        "type": "number"
                    },
                    "daily_history": {
                        "items": {
                            "$ref": "#/components/schemas/DailyPoint"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "electrical": {
                        "$ref": "#/components/schemas/Electrical"
                    },
                    "events": {
                        "description": "eventos operativos registrados para el medidor",
                        "items": {
                            "$ref": "#/components/schemas/Event"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "hourly_history": {
                        "items": {
                            "$ref": "#/components/schemas/HourlyPoint"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "last_reading_at": {
                        "type": "string"
                    },
                    "meter_id": {
                        "example": "M-109",
                        "type": "string"
                    },
                    "priority": {
                        "description": "1 = revisar primero; null si no hay anomalía",
                        "example": 1,
                        "type": "integer"
                    },
                    "severity": {
                        "$ref": "#/components/schemas/Severity",
                        "description": "null si no hay anomalía"
                    },
                    "status": {
                        "$ref": "#/components/schemas/Status"
                    },
                    "variation_pct": {
                        "description": "consumo vs. baseline",
                        "example": 109.82,
                        "type": "number"
                    }
                },
                "type": "object"
            },
            "Direction": {
                "enum": [
                    "UP",
                    "DOWN"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "Up",
                    "Down"
                ]
            },
            "Electrical": {
                "properties": {
                    "current_a": {
                        "$ref": "#/components/schemas/Metric"
                    },
                    "power_factor": {
                        "$ref": "#/components/schemas/Metric"
                    },
                    "voltage_v": {
                        "$ref": "#/components/schemas/Metric"
                    }
                },
                "type": "object"
            },
            "Event": {
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "meter_id": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string"
                    },
                    "type": {
                        "$ref": "#/components/schemas/EventType"
                    }
                },
                "type": "object"
            },
            "EventType": {
                "enum": [
                    "OPERATIONAL_CHANGE",
                    "SCHEDULED_OUTAGE",
                    "DATA_QUALITY",
                    "UNKNOWN"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "EventOperationalChange",
                    "EventScheduledOutage",
                    "EventDataQuality",
                    "EventUnknown"
                ]
            },
            "Evidence": {
                "properties": {
                    "baseline_daily_kwh": {
                        "type": "number"
                    },
                    "change": {
                        "$ref": "#/components/schemas/ConsumptionChange"
                    },
                    "checks": {
                        "items": {
                            "$ref": "#/components/schemas/Check"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "daily_change_pct": {
                        "type": "number"
                    },
                    "impact_kwh": {
                        "description": "energía de más (+) o de menos (-) frente a lo esperado",
                        "type": "number"
                    },
                    "last_day_kwh": {
                        "type": "number"
                    },
                    "quality": {
                        "$ref": "#/components/schemas/Report"
                    },
                    "related_events": {
                        "items": {
                            "$ref": "#/components/schemas/Event"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "variables": {
                        "items": {
                            "$ref": "#/components/schemas/VariableChange"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "window_end": {
                        "type": "string"
                    },
                    "window_start": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Filter": {
                "properties": {
                    "meter_id": {
                        "description": "Coincidencia parcial y sin importar mayúsculas: \"109\" encuentra M-109.",
                        "example": "M-109",
                        "type": "string"
                    },
                    "sort_by": {
                        "$ref": "#/components/schemas/SortField",
                        "default": "meter_id"
                    },
                    "sort_order": {
                        "$ref": "#/components/schemas/SortOrder",
                        "default": "ASC"
                    },
                    "status": {
                        "$ref": "#/components/schemas/Status"
                    }
                },
                "type": "object"
            },
            "HourlyPoint": {
                "properties": {
                    "consumption_kwh": {
                        "example": 110.35,
                        "type": "number"
                    },
                    "current_a": {
                        "example": 411.1,
                        "type": "number"
                    },
                    "expected_kwh": {
                        "description": "baseline de esa hora del día",
                        "example": 51.66,
                        "type": "number"
                    },
                    "power_factor": {
                        "example": 0.74,
                        "type": "number"
                    },
                    "timestamp": {
                        "type": "string"
                    },
                    "voltage_v": {
                        "example": 218.4,
                        "type": "number"
                    }
                },
                "type": "object"
            },
...
            "MeterSummary": {
                "properties": {
                    "anomaly_type": {
                        "$ref": "#/components/schemas/AnomalyType"
                    },
                    "baseline_kwh": {
                        "description": "consumo diario normal",
                        "example": 1052.15,
                        "type": "number"
                    },
                    "consumption_kwh": {
                        "description": "consumo de las últimas 24 h",
                        "example": 2207.6,
                        "type": "number"
                    },
                    "last_reading_at": {
                        "type": "string"
                    },
                    "meter_id": {
                        "example": "M-109",
                        "type": "string"
                    },
                    "priority": {
                        "description": "1 = revisar primero; null si no hay anomalía",
                        "example": 1,
                        "type": "integer"
                    },
                    "severity": {
                        "$ref": "#/components/schemas/Severity",
                        "description": "null si no hay anomalía"
                    },
                    "status": {
                        "$ref": "#/components/schemas/Status"
                    },
                    "variation_pct": {
                        "description": "consumo vs. baseline",
                        "example": 109.82,
                        "type": "number"
                    }
                },
                "type": "object"
            },
            "Metric": {
                "properties": {
                    "baseline": {
                        "description": "mediana de la semana de referencia",
                        "example": 219.7,
                        "type": "number"
                    },
                    "change_pct": {
                        "example": -0.59,
                        "type": "number"
                    },
                    "value": {
                        "description": "última lectura",
                        "example": 218.4,
                        "type": "number"
                    }
                },
                "type": "object"
            },
            "Option": {
                "properties": {
                    "id": {
                        "example": "CRITICAL",
                        "type": "string"
                    },
                    "value": {
                        "example": "Crítica",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "PagedRequest": {
                "properties": {
                    "filter": {
                        "$ref": "#/components/schemas/Filter"
                    },
                    "pagination": {
                        "$ref": "#/components/schemas/Pagination"
                    }
                },
                "type": "object"
            },
            "PaginatedResult-MeterSummary": {
                "properties": {
                    "count": {
                        "description": "total de resultados, sin paginar",
                        "example": 12,
                        "type": "integer"
                    },
                    "page": {
                        "example": 1,
                        "type": "integer"
                    },
                    "rows": {
                        "items": {
                            "$ref": "#/components/schemas/MeterSummary"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "size": {
                        "example": 10,
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Pagination": {
                "properties": {
                    "page": {
                        "example": 1,
                        "minimum": 1,
                        "type": "integer"
                    },
                    "size": {
                        "example": 10,
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "ParamsResponse": {
                "properties": {
                    "meters": {
                        "description": "un ítem por medidor: { id: \"M-109\", value: \"M-109\" }",
                        "items": {
                            "$ref": "#/components/schemas/Option"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "sort_fields": {
                        "description": "meter_id, consumption, variation, severity",
                        "items": {
                            "$ref": "#/components/schemas/Option"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "sort_orders": {
                        "description": "ASC, DESC",
                        "items": {
                            "$ref": "#/components/schemas/Option"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "statuses": {
                        "description": "NORMAL, ALERT, CRITICAL",
                        "items": {
                            "$ref": "#/components/schemas/Option"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
            "Report": {
                "properties": {
                    "duplicate_readings": {
                        "description": "dos lecturas con la misma hora",
                        "type": "integer"
                    },
                    "first_flagged": {
                        "type": "string"
                    },
                    "flagged_hours": {
                        "description": "horas con al menos un problema",
                        "type": "integer"
                    },
                    "invalid_values": {
                        "description": "valores imposibles (FP \u003e 1, negativos...)",
                        "type": "integer"
                    },
                    "last_flagged": {
                        "type": "string"
                    },
                    "missing_hours": {
                        "description": "horas sin lectura",
                        "type": "integer"
                    },
                    "power_factor_mismatch": {
                        "description": "FP disparado con consumo normal",
                        "type": "integer"
                    },
                    "voltage_out_of_range": {
                        "description": "voltaje a más de ±5% de lo normal",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "Severity": {
                "enum": [
                    "HIGH",
                    "MEDIUM",
                    "LOW"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "SeverityHigh",
                    "SeverityMedium",
                    "SeverityLow"
                ]
            },
            "SortField": {
                "enum": [
                    "meter_id",
                    "consumption",
                    "variation",
                    "severity"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "SortByMeterID",
                    "SortByConsumption",
                    "SortByVariation",
                    "SortBySeverity"
                ]
            },
            "SortOrder": {
                "enum": [
                    "ASC",
                    "DESC"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "Asc",
                    "Desc"
                ]
            },
            "Status": {
                "enum": [
                    "NORMAL",
                    "ALERT",
                    "CRITICAL"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "StatusNormal",
                    "StatusAlert",
                    "StatusCritical"
                ]
            },
...
            "VariableChange": {
                "properties": {
                    "baseline": {
                        "description": "mediana normal, en las mismas horas del día",
                        "type": "number"
                    },
                    "change_pct": {
                        "type": "number"
                    },
                    "changed": {
                        "type": "boolean"
                    },
                    "max": {
                        "type": "number"
                    },
                    "min": {
                        "type": "number"
                    },
                    "observed": {
                        "description": "mediana durante el problema",
                        "type": "number"
                    },
                    "variable": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "bearerFormat": "JWT",
                "description": "Pega solo el token (sin \"Bearer\").",
                "scheme": "bearer",
                "type": "http"
            }
        }
    },
    "info": {
        "description": "API para monitorear medidores eléctricos y priorizar sus anomalías.\nPrimero haz login en POST /auth/login, copia el `access_token` y pégalo en **Authorize**.",
        "title": "EnergIA API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/meter/getAll": {
            "post": {
                "description": "Filtros opcionales: `meter_id` (parcial), `status` (NORMAL, ALERT, CRITICAL).\nOrden: `sort_by` (meter_id, consumption, variation, severity) y `sort_order` (ASC, DESC).\nSin body devuelve la página 1 de 10, ordenada por medidor.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PagedRequest",
                                "summary": "body",
                                "description": "Paginación y filtros"
                            }
                        }
                    },
                    "description": "Paginación y filtros"
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedResult-MeterSummary"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIError"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIError"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Lista de medidores, filtrada, ordenada y paginada",
                "tags": [
                    "Meters"
                ]
            }
        },
        "/meter/getById/{id}": {
            "get": {
                "description": "Consumo actual (últimas 24 h), baseline, variación, estado, voltaje, corriente,\nfactor de potencia, histórico horario y diario, anomalía y eventos.",
                "parameters": [
                    {
                        "description": "ID del medidor",
                        "example": "M-109",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Detail"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIError"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIError"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Detalle de un medidor",
                "tags": [
                    "Meters"
                ]
            }
        },
        "/meter/getParams": {
            "get": {
                "description": "Opciones de los desplegables: medidores, estados de alerta, campos y dirección de ordenamiento.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ParamsResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/APIError"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Opciones de los filtros",
                "tags": [
                    "Meters"
                ]
            }
        }
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/api/v1"
        }
    ]
}